Fix issue saving configuration
CascadingConfigParser.save loads all config files and checks for
differences to the currently loaded config. If the changes are to be
saved to one of the config files (local.cfg by default) then this file
should not be loaded, otherwise any content of this file will be lost,
i.e. any item that is configured in local.cfg would be loaded and
compared to the current config, if it matches, it would not be stored
back into local.cfg.
This change fixes this by removing the file from the list of config
files to be loaded.